fix: cross-widget performance & correctness pass (0.2.10)
Full audit of all widgets (Dart + native Swift):
- Remove native->Dart->native echo loops in slider/toggle/segmented
- Add mounted guards + safe casts across native-call handlers
- Cache decoded icon images in button/toolbar; stable SwiftUI identity
- Push stale state to native (menu items, scaffold selectedIndex,
  search bar controls on iOS 26 SwiftUI path, nav bar setStyle reset)
- Fix shared presenter channel handler clobbering (alert/sheet/popover)
  and release user-dismissed sheets via presentation delegate
- Debounce search text; guard SwiftUI slider range; clamp segmented
  selection; stepper range resync; container listEquals + memoized
  params; Int64 timestamp; colorWell size; progress animation

Verified: flutter analyze clean; flutter build ios --simulator compiles
all plugin Swift. Runtime behavior pending device testing.

+++ ios/native_liquid_glass/Sources/native_liquid_glass/LiquidGlassDatePicker/LiquidGlassDatePickerView.swift
@@ -103,7 +103,7 @@ final class LiquidGlassDatePickerPlatformView: NSObject, FlutterPlatformView {
 
   @objc
   private func handleDateChanged() {
-    let millis = Int(datePicker.date.timeIntervalSince1970 * 1000)
+    let millis = Int64(datePicker.date.timeIntervalSince1970 * 1000)
     methodChannel.invokeMethod("dateChanged", arguments: millis)
   }
